velvet-radio/requirements.txt
# ============================================================
#  Velvet Radio — Python 의존성
# ============================================================

# Core
httpx>=0.27
pydantic>=2.0
pydantic-settings>=2.0
python-dotenv>=1.0
tenacity>=8.0

# Claude API
anthropic>=0.40

# Audio Analysis
librosa>=0.10
pyloudnorm>=0.1
numpy>=1.24

# YouTube / Google APIs
google-api-python-client>=2.120
google-auth-oauthlib>=1.2
google-auth-httplib2>=0.2

# Image Processing (fallback)
Pillow>=10.0

# Logging
structlog>=24.0

# Dashboard (Web UI)
Flask>=3.0
Werkzeug>=3.0

# Testing
pytest>=8.0
pytest-asyncio>=0.23
pytest-mock>=3.12
respx>=0.21  # httpx mock

# Linting & Formatting (dev)
ruff>=0.3
mypy>=1.9
pre-commit>=3.7
